+++ src/config/common.php
@@ -8,7 +8,9 @@
 return [
     "name"          => "SkeekS CMS",
     'id'            => 'skeeks-cms-app',
+
     'vendorPath'    => '@vendor',
+
     'language'      => 'ru',
 
     'components' => [